move keys to env file
