+++ src/Plugin/firebase.js
@@ -1,12 +1,15 @@
-import * as firebase from 'firebase';
+// import * as firebase from 'firebase';
+// import "firebase/firestore";
+// const firebaseApp = firebase.initializeApp(config);
 
-var config = {
-  apiKey: process.env.VUE_APP_FIREBASE_API_KEY,
-  authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-  databaseURL: process.env.FIREBASE_DATABASE_URL,
-  projectId: "shecluded",
-  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.FIREBASE_MESSANGER_SENDER_ID,
-  appId: process.env.FIREBASE_APP_ID
-};
-export default !firebase.apps.length ? firebase.initializeApp(config) : firebase;
+// var config = {
+//   apiKey: process.env.VUE_APP_FIREBASE_API_KEY,
+//   authDomain: process.env.FIREBASE_AUTH_DOMAIN,
+//   databaseURL: process.env.FIREBASE_DATABASE_URL,
+//   projectId: "shecluded",
+//   storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
+//   messagingSenderId: process.env.FIREBASE_MESSANGER_SENDER_ID,
+//   appId: process.env.FIREBASE_APP_ID
+// };
+
+// export default firebaseApp.firestore();
