--- src/Plugin/firebaseConfig.js
@@ -1,11 +1,0 @@
-var firebaseConfig = {
-  apiKey: "AIzaSyAyJHKm9_avmoPADfiX8hIp4dcTCys0t0Y",
-  authDomain: "shecluded.firebaseapp.com",
-  databaseURL: "https://shecluded.firebaseio.com",
-  projectId: "shecluded",
-  storageBucket: "shecluded.appspot.com",
-  messagingSenderId: "795277390492",
-  appId: "1:795277390492:web:4ce2dd5342e698b313b359",
-  measurementId: "G-KQV1WRF42W"
- };
- export default firebaseConfig;
