+++ src/firebase.js
@@ -1,7 +1,15 @@
-import * as firebase from "firebase";
+import * as firebase from 'firebase';
 import "firebase/firestore";
-import firebaseConfig from "./Plugin/firebaseConfig";
 
-const firebaseApp = firebase.initializeApp(firebaseConfig);
+var config = {
+  apiKey: process.env.VUE_APP_FIREBASE_API_KEY,
+  authDomain: process.env.FIREBASE_AUTH_DOMAIN,
+  databaseURL: process.env.FIREBASE_DATABASE_URL,
+  projectId: "shecluded",
+  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: process.env.FIREBASE_MESSANGER_SENDER_ID,
+  appId: process.env.FIREBASE_APP_ID
+};
+const firebaseApp = firebase.initializeApp(config);
 
 export default firebaseApp.firestore();
